// src/nftmusic/dto/info.nfthistory.dto.ts
export class InfoNftHistoryDto {

  id: number;

  symbol: string;

  tokenId: string;

  price: string;

  fromAddress: string;

  toAddress: string;

  timestamp: string;

  txHash: string;

  status: string;

  blockNumber: string;

  method: string;

  extra: string;

  name: string;

  url: string;
}